+++ packages/terra-form-checkbox/src/Checkbox.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Checkbox.scss';
+import styles from './Checkbox.module.scss';
 import CheckboxUtil from './CheckboxUtil';
 
 const cx = classNames.bind(styles);