+++ packages/terra-form-input/src/Input.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Input.scss';
+import styles from './Input.module.scss';
 
 const cx = classNames.bind(styles);