+++ packages/terra-hyperlink/src/Hyperlink.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Hyperlink.scss';
+import styles from './Hyperlink.module.scss';
 
 const cx = classNames.bind(styles);