+++ packages/terra-legacy-theme/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Add .module extension to test SCSS file(s)
 
 2.17.0 - (July 17, 2018)
 ------------------